requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.15
vk-api==11.9.9
